+++ packages/perseus-score/src/widgets/mock-widget/mock-widget-validation.types.ts
@@ -1,0 +1,7 @@
+export type PerseusMockWidgetRubric = {
+    value: string;
+};
+
+export type PerseusMockWidgetUserInput = {
+    currentValue: string;
+};
