+++ packages/perseus/src/__tests__/server-item-renderer.test.tsx
@@ -66,8 +66,6 @@ const renderQuestion = (
 
 describe("server item renderer", () => {
     beforeAll(() => {
-        // TODO(LEMS-2656): remove TS suppression
-        // @ts-expect-error: MockWidget is not assignable to type WidgetExports
         registerWidget("mock-widget", MockWidget);
     });
