+++ packages/perseus/src/widget-ai-utils/mock-widget/prompt-utils.test.ts
@@ -1,6 +1,6 @@
 import {getPromptJSON} from "./prompt-utils";
 
-import type {PerseusMockWidgetUserInput} from "../../widgets/mock-widgets/mock-widget-types";
+import type {PerseusMockWidgetUserInput} from "@khanacademy/perseus-score";
 
 describe("InputNumber getPromptJSON", () => {
     it("it returns JSON with the expected format and fields", () => {
